Abstract the editor's positioning logic
