+++ handsontable/src/3rdparty/walkontable/src/overlay/top.js
@@ -58,6 +58,30 @@ export class TopOverlay extends Overlay {
     return this.wtSettings.getSetting('shouldRenderTopOverlay');
   }
 
+  getOverlayPosition() {
+    const overlayRoot = this.clone.wtTable.holder.parentNode;
+    const { rootWindow } = this.domBindings;
+    const preventOverflow = this.wtSettings.getSetting('preventOverflow');
+    let y = 0;
+
+    if (this.trimmingContainer === rootWindow && (!preventOverflow || preventOverflow !== 'vertical')) {
+      const { wtTable } = this.wot;
+      const hiderRect = wtTable.hider.getBoundingClientRect();
+      const top = Math.ceil(hiderRect.top);
+      const bottom = Math.ceil(hiderRect.bottom);
+      const rootHeight = overlayRoot.offsetHeight;
+
+      if (top < 0 && (bottom - rootHeight) > 0) {
+        y = -top;
+      }
+    }
+
+    return {
+      x: 0,
+      y,
+    };
+  }
+
   /**
    * Updates the top overlay position.
    *
@@ -69,16 +93,15 @@ export class TopOverlay extends Overlay {
       return false;
     }
 
-    const { rootWindow } = this.domBindings;
     const overlayRoot = this.clone.wtTable.holder.parentNode;
+    const { rootWindow } = this.domBindings;
     const preventOverflow = this.wtSettings.getSetting('preventOverflow');
-    let headerPosition = 0;
+    let overlayPosition = 0;
     let skipInnerBorderAdjusting = false;
 
     if (this.trimmingContainer === rootWindow && (!preventOverflow || preventOverflow !== 'vertical')) {
       const { wtTable } = this.wot;
       const hiderRect = wtTable.hider.getBoundingClientRect();
-      const top = Math.ceil(hiderRect.top);
       const bottom = Math.ceil(hiderRect.bottom);
       const rootHeight = overlayRoot.offsetHeight;
 
@@ -94,22 +117,16 @@ export class TopOverlay extends Overlay {
       // This workaround will be able to be cleared after merging the SVG borders, which introduces
       // frozen lines (no more `innerBorderTop` workaround).
       skipInnerBorderAdjusting = bottom === rootHeight;
+      overlayPosition = this.getOverlayPosition().y;
 
-      let finalTop = 0;
-
-      if (top < 0 && (bottom - rootHeight) > 0) {
-        finalTop = -top;
-      }
-
-      headerPosition = finalTop;
-      setOverlayPosition(overlayRoot, '0px', `${finalTop}px`);
+      setOverlayPosition(overlayRoot, '0px', `${overlayPosition}px`);
 
     } else {
-      headerPosition = this.getScrollPosition();
+      overlayPosition = this.getScrollPosition();
       resetCssTransform(overlayRoot);
     }
 
-    const positionChanged = this.adjustHeaderBordersPosition(headerPosition, skipInnerBorderAdjusting);
+    const positionChanged = this.adjustHeaderBordersPosition(overlayPosition, skipInnerBorderAdjusting);
 
     this.adjustElementsSize();
