+++ handsontable/src/editors/textEditor/textEditor.js
@@ -6,16 +6,10 @@ import {
   getCaretPosition,
   getComputedStyle,
   getCssTransform,
-  getScrollbarWidth,
-  innerWidth,
-  offset,
   resetCssTransform,
   setCaretPosition,
-  hasVerticalScrollbar,
-  hasHorizontalScrollbar,
   hasClass,
   removeClass,
-  outerWidth
 } from '../../helpers/dom/element';
 import { stopImmediatePropagation, isImmediatePropagationStopped } from '../../helpers/dom/event';
 import { rangeEach } from '../../helpers/number';
@@ -339,124 +333,24 @@ export class TextEditor extends BaseEditor {
       return;
     }
 
-    const { wtOverlays, wtViewport } = this.hot.view._wt;
-    const rootWindow = this.hot.rootWindow;
-    const currentOffset = offset(this.TD);
-    const cellWidth = outerWidth(this.TD);
-    const containerOffset = offset(this.hot.rootElement);
-    const containerWidth = outerWidth(this.hot.rootElement);
-    const scrollableContainerTop = wtOverlays.topOverlay.holder;
-    const scrollableContainerLeft = wtOverlays.inlineStartOverlay.holder;
-    const containerScrollTop = scrollableContainerTop !== rootWindow ?
-      scrollableContainerTop.scrollTop : 0;
-    const containerScrollLeft = scrollableContainerLeft !== rootWindow ?
-      scrollableContainerLeft.scrollLeft : 0;
-    const gridMostRightPos = rootWindow.innerWidth - containerOffset.left - containerWidth;
-    const editorSection = this.checkEditorSection();
+    const {
+      top,
+      start,
+      width,
+      maxWidth,
+      height,
+      maxHeight
+    } = this.getEditedCellRect();
 
-    const scrollTop = ['', 'inline-start'].includes(editorSection) ? containerScrollTop : 0;
-    const scrollLeft = ['', 'top', 'bottom'].includes(editorSection) ? containerScrollLeft : 0;
-
-    // If colHeaders is disabled, cells in the first row have border-top
-    const editTopModifier = currentOffset.top === containerOffset.top ? 0 : 1;
-    const backgroundColor = this.TD.style.backgroundColor;
-
-    let editTop = currentOffset.top - containerOffset.top - editTopModifier - scrollTop;
-    let inlineStartPos = 0;
-
-    if (this.hot.isRtl()) {
-      inlineStartPos = rootWindow.innerWidth - currentOffset.left - cellWidth - gridMostRightPos - 1 + scrollLeft;
-    } else {
-      inlineStartPos = currentOffset.left - containerOffset.left - 1 - scrollLeft;
-    }
-
-    let wtTable = this.hot.view._wt.wtTable;
-
-    // TODO: Refactor this to the new instance.getCell method (from #ply-59), after 0.12.1 is released
-    switch (editorSection) {
-      case 'top':
-        wtTable = wtOverlays.topOverlay.clone.wtTable;
-        break;
-      case 'inline-start':
-        wtTable = wtOverlays.inlineStartOverlay.clone.wtTable;
-        break;
-      case 'top-inline-start-corner':
-        wtTable = wtOverlays.topInlineStartCornerOverlay.clone.wtTable;
-        break;
-      case 'bottom-inline-start-corner':
-        wtTable = wtOverlays.bottomInlineStartCornerOverlay.clone.wtTable;
-        break;
-      case 'bottom':
-        wtTable = wtOverlays.bottomOverlay.clone.wtTable;
-        break;
-      default:
-        break;
-    }
-
-    const cssTransformOffset = getCssTransform(wtTable.holder.parentNode);
-    const hasColumnHeaders = this.hot.hasColHeaders();
-    const renderableRow = this.hot.rowIndexMapper.getRenderableFromVisualIndex(this.row);
-    const renderableColumn = this.hot.columnIndexMapper.getRenderableFromVisualIndex(this.col);
-    const nrOfRenderableRowIndexes = this.hot.rowIndexMapper.getRenderableIndexesLength();
-    const firstRowIndexOfTheBottomOverlay = nrOfRenderableRowIndexes - this.hot.view._wt.getSetting('fixedRowsBottom');
-
-    if (hasColumnHeaders && renderableRow <= 0 || renderableRow === firstRowIndexOfTheBottomOverlay) {
-      editTop += 1;
-    }
-
-    if (renderableColumn <= 0) {
-      inlineStartPos += 1;
-    }
-
-    if (cssTransformOffset && cssTransformOffset !== -1) {
-      this.textareaParentStyle[cssTransformOffset[0]] = cssTransformOffset[1];
-    } else {
-      resetCssTransform(this.TEXTAREA_PARENT);
-    }
-
-    this.textareaParentStyle.top = `${editTop}px`;
-    this.textareaParentStyle[this.hot.isRtl() ? 'right' : 'left'] = `${inlineStartPos}px`;
+    this.textareaParentStyle.top = `${top}px`;
+    this.textareaParentStyle[this.hot.isRtl() ? 'right' : 'left'] = `${start}px`;
     this.showEditableElement();
-
-    const firstRowOffset = wtViewport.rowsRenderCalculator.startPosition;
-    const firstColumnOffset = wtViewport.columnsRenderCalculator.startPosition;
-    const horizontalScrollPosition = Math.abs(wtOverlays.inlineStartOverlay.getScrollPosition());
-    const verticalScrollPosition = wtOverlays.topOverlay.getScrollPosition();
-    const scrollbarWidth = getScrollbarWidth(this.hot.rootDocument);
-
-    const cellTopOffset = this.TD.offsetTop + firstRowOffset - verticalScrollPosition;
-
-    let cellStartOffset = 0;
-
-    if (this.hot.isRtl()) {
-      const cellOffset = this.TD.offsetLeft;
-
-      if (cellOffset >= 0) {
-        cellStartOffset = wtTable.getWidth() - this.TD.offsetLeft;
-      } else {
-        // The `offsetLeft` returns negative values when the parent offset element has position relative
-        // (it happens when on the cell the selection is applied - the `area` CSS class).
-        // When it happens the `offsetLeft` value is calculated from the right edge of the parent element.
-        cellStartOffset = Math.abs(cellOffset);
-      }
-
-      cellStartOffset += firstColumnOffset - horizontalScrollPosition - cellWidth;
-    } else {
-      cellStartOffset = this.TD.offsetLeft + firstColumnOffset - horizontalScrollPosition;
-    }
-
-    const actualVerticalScrollbarWidth = hasVerticalScrollbar(scrollableContainerTop) ? scrollbarWidth : 0;
-    const actualHorizontalScrollbarWidth = hasHorizontalScrollbar(scrollableContainerLeft) ? scrollbarWidth : 0;
-    const width = innerWidth(this.TD) - 8;
-    const maxWidth = this.hot.view.maximumVisibleElementWidth(cellStartOffset) - 9 - actualVerticalScrollbarWidth;
-    const height = this.TD.scrollHeight + 1;
-    const maxHeight = Math.max(this.hot.view.maximumVisibleElementHeight(cellTopOffset) - actualHorizontalScrollbarWidth, 23); // eslint-disable-line max-len
 
     const cellComputedStyle = getComputedStyle(this.TD, this.hot.rootWindow);
 
     this.TEXTAREA.style.fontSize = cellComputedStyle.fontSize;
     this.TEXTAREA.style.fontFamily = cellComputedStyle.fontFamily;
-    this.TEXTAREA.style.backgroundColor = backgroundColor;
+    this.TEXTAREA.style.backgroundColor = this.TD.style.backgroundColor;
 
     this.autoResize.init(this.TEXTAREA, {
       minHeight: Math.min(height, maxHeight),
